🔀 refactor(defaults.ts): remove unnecessary code for setting default values 🔀 refactor(constants.ts): change default value for enable_automatically_set_quality to false
The code in defaults.ts was setting an unnecessary additional key-value pair in the local storage. This code has been removed to simplify the logic.

In constants.ts, the default value for enable_automatically_set_quality has been changed to false. This means that the feature to automatically set the quality of the player is now disabled by default.