feat(Settings.tsx): add functionality has been added to automatically save settings
:twisted_rightwards_arrows: refactor(Settings.css): update padding values for .notification class
:twisted_rightwards_arrows: refactor(SettingsWrapper/index.tsx): update defaultOptions assignment

The padding values for the `.notification` class in `Settings.css` have been updated to `10px` for top and bottom and `14px` for left and right. This change improves the visual appearance of the notifications.

In `Settings.tsx`, functionality has been added to automatically save settings. When a checkbox or input value is changed, the corresponding setting is updated.

In `SettingsWrapper/index.tsx`, the assignment of `defaultOptions` has been updated to use the `defaultConfiguration` constant instead of dynamically generating it from `localStorage`. This change ensures that the default options are always consistent and avoids potential issues with missing or incorrect default values.