End screen cards flag fix
Put the class for hiding end screen cards on the `div#movie_player` and
specify in the css to hide the relevant descendants. Also including to
hide the ending endscreen as I assume that for most people, those "cards"
are "end screen cards" and not some other "watch more crumbs" or
similar.

The reason for putting the flag on the top class is that the actual end
screen cards don't always have time to be loaded before the relevant
code is executed and thus won't be hidden.